timescale-dataviz/pyproject.toml
[tool.poetry]
name = "timescale-langchain"
version = "0.1.0"
description = ""
authors = ["Jolkkonen, Johannes (External) <johannes.jolkkonen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-community = "^0.0.36"
langchain = "^0.1.17"
timescale-vector = "^0.0.4"
python-decouple = "^3.8"
langchain-openai = "^0.1.6"
pandas = "^2.2.2"
lark = "^1.1.9"
instructor = "^1.2.5"
openai = "^1.26.0"
anthropic = "^0.25.8"
streamlit = "^1.34.0"
plotly = "^5.22.0"
black = "^24.4.2"
groq = "^0.5.0"
statsmodels = "^0.14.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
